+++ api/src/main/java/org/openmrs/api/db/hibernate/HibernateEncounterDAO.java
@@ -97,9 +97,9 @@ public class HibernateEncounterDAO implements EncounterDAO {
 	 */
 	@SuppressWarnings("unchecked")
 	public List<Encounter> getEncountersByPatientId(Integer patientId) throws DAOException {
-		Criteria crit = sessionFactory.getCurrentSession().createCriteria(Encounter.class).createAlias("patient", "p")
-		        .add(Expression.eq("p.patientId", patientId)).add(Expression.eq("voided", false))
-		        .addOrder(Order.desc("encounterDatetime"));
+		Criteria crit = sessionFactory.getCurrentSession().createCriteria(Encounter.class).createAlias("patient", "p").add(
+		    Expression.eq("p.patientId", patientId)).add(Expression.eq("voided", false)).addOrder(
+		    Order.desc("encounterDatetime"));
 		
 		return crit.list();
 	}
@@ -207,9 +207,9 @@ public class HibernateEncounterDAO implements EncounterDAO {
 	@SuppressWarnings("unchecked")
 	public List<EncounterType> findEncounterTypes(String name) throws DAOException {
 		return sessionFactory.getCurrentSession().createCriteria(EncounterType.class)
-		        // 'ilike' case insensitive search
-		        .add(Expression.ilike("name", name, MatchMode.START)).addOrder(Order.asc("name"))
-		        .addOrder(Order.asc("retired")).list();
+		// 'ilike' case insensitive search
+		        .add(Expression.ilike("name", name, MatchMode.START)).addOrder(Order.asc("name")).addOrder(
+		            Order.asc("retired")).list();
 	}
 	
 	/**
@@ -315,11 +315,11 @@ public class HibernateEncounterDAO implements EncounterDAO {
 	 * @see org.openmrs.api.db.EncounterDAO#getCountOfEncounters(java.lang.String, boolean)
 	 */
 	@Override
-	public Integer getCountOfEncounters(String query, boolean includeVoided) {
+	public Long getCountOfEncounters(String query, boolean includeVoided) {
 		Criteria criteria = createEncounterByQueryCriteria(query, includeVoided);
 		
 		criteria.setProjection(Projections.countDistinct("enc.encounterId"));
-		return (Integer) criteria.uniqueResult();
+		return (Long) criteria.uniqueResult();
 	}
 	
 	/**
@@ -356,8 +356,8 @@ public class HibernateEncounterDAO implements EncounterDAO {
 	@SuppressWarnings("unchecked")
 	public List<Encounter> getEncountersByVisit(Visit visit) {
 		Criteria crit = sessionFactory.getCurrentSession().createCriteria(Encounter.class).createAlias("visit", "visit")
-		        .add(Expression.eq("visit", visit)).add(Expression.eq("voided", false))
-		        .addOrder(Order.asc("encounterDatetime"));
+		        .add(Expression.eq("visit", visit)).add(Expression.eq("voided", false)).addOrder(
+		            Order.asc("encounterDatetime"));
 		
 		return crit.list();
 	}
